requirements.txt
python-telegram-bot==20.0
schedule==1.1.0
httpx==0.23.3
